Safety backup, got rid of .swp, more changes to (still not working) new network, separated files (use the same one as before, 0 is for keeping changes)
